fix(python-sdk): keep codegen-owned envd __init__ and fix stale test mock
The Generated files check regenerates e2b/envd/__init__.py (protoc-gen-py
writes it), so the shared codec/proxy/pool-tuning plumbing moves to
e2b/envd/client_shared.py and __init__.py returns to its generated
content. Also updates the thread-local clients test mock that still
declared the removed sync keyword argument.
